pyproject.toml
[project]
name = "murray-munchers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipython>=9.8.0",
    "numpy>=2.4.0",
    "pygame-ce>=2.5.6",
    "pygame-menu>=4.5.2",
    "pyinstaller>=6.17.0",
    "ruff>=0.14.10",
]
